Better yet, I've moved the PhysicalConstants class to a new file called sandbox in Utilities.  This can be used as follows:
import OpenPNM.Utilities.sandbox as sandbox
air.physical_constants = sandbox.PhysicalConstants()
