+++ OpenPNM/Utilities/sandbox.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+
+from collections import _namedtuple
+
+
+class PhysicalConstants:
+    r"""
+    A generic object contain the intrinsic properties of the Phase
+    """
+    molecular_weight = None
+    critical_pressure = None
+    critical_temperature = None
+    melting_point = None
+    freezing_point = None
+    antoines_coefficients = _namedtuple('antoines_coefficients',
+                                        ('A', 'B', 'C'))
